Extract pure App helpers and defaults
